+++ NEWS.rst
@@ -6,8 +6,9 @@ API Additions
 
 * ``rouver.test``: Add ``TestRequest.set_env_var()`` and
   ``TestRequest.set_header()``.
-* ``rouver.test``: Add ``TestResponse.assert_content_type()``,
-  ``assert_see_other()``, and ``assert_temporary_redirect()``.
+* ``rouver.test``: Add ``TestResponse.assert_header_not_set()``,
+  ``assert_content_type()``, ``assert_see_other()``, and
+  ``assert_temporary_redirect()``.
 
 News in rouver 0.10.0
 =====================